+++ database/migrations/2019_09_02_221103_add_extra_cols_to_users.php
@@ -17,7 +17,7 @@ class AddExtraColsToUsers extends Migration
             $table->string('phone')->after('email');
             $table->string('alt_phone')->after('phone')->nullable();
             $table->string('address')->after('alt_phone')->nullable();
-            $table->integer('location_id')->after('address')->nullable();
+            $table->bigInteger('location_id')->after('address')->nullable();
             $table->string('area_code')->after('location_id')->nullable();
             $table->boolean('status')->after('area_code')->default(1);
         });
