+++ database/migrations/2019_09_11_121146_drop-user_id_col_in_messages.php
@@ -15,7 +15,7 @@ class DropUserIdColInMessages extends Migration
     {
         Schema::table('messages', function (Blueprint $table) {
             $table->dropColumn('user_id');
-            $table->integer('sender_id')->unsigned();
+            $table->bigInteger('sender_id')->unsigned();
 
             $table->foreign('sender_id')->references('id')->on('users');
         });
